Add version scripts for ESLint packages
New `git-tag-version` script that prepends the package name to the version number so that we can
have unique tags across the entire repo. Also commits the package.json since `npm version`
no longer does that.

Only Eventbrite OPS has access to publish to NPM, so the scripts just version the package and
push tags to git.